Refactor : update types
