+++ src/types/User.ts
@@ -1,0 +1,5 @@
+interface UserData {
+  bull_tokens: number;
+  bear_tokens: number;
+}
+export type { UserData };
